rename stocks service directory
